split api and client
